fix:
- Explicitly declare variables for readability
- Removed unreferenced method List<KeyValuePair<string, string>> ParseKeyed(IEnumerable<string> input)
- Move command filter method to separate class to keep controller clear
