+++ toolkitcore/ToolkitCore.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ToolkitCoreSettings.cs" />
     <Compile Include="TwitchInterfaceBase.cs" />
+    <Compile Include="Utilities\CommandFilter.cs" />
     <Compile Include="Utilities\TCText.cs" />
     <Compile Include="Utilities\ViewerInterface.cs" />
   </ItemGroup>
